Multiple modules: clean up unit tests (#11630)
* Test chunking

* update some bidder eid tests

* split eid tests into each userId submodule

* cleanup userId_spec

* add TEST_PAT config

* fix idx, lmp

* clean up userId_spec

* fix double run, invibes, intentIq

* small fixes

* undo package-lock changes

* update colors, remove empty test

* 8pod analytics: clean up interval handler
