+++ karma.conf.maker.js
@@ -108,13 +108,11 @@ function setBrowsers(karmaConf, browserstack) {
 module.exports = function(codeCoverage, browserstack, watchMode, file, disableFeatures) {
   var webpackConfig = newWebpackConfig(codeCoverage, disableFeatures);
   var plugins = newPluginsArray(browserstack);
-
-  var files = file ? ['test/test_deps.js', file, 'test/helpers/hookSetup.js'].flatMap(f => f) : ['test/test_index.js'];
-  // This file opens the /debug.html tab automatically.
-  // It has no real value unless you're running --watch, and intend to do some debugging in the browser.
-  if (watchMode) {
-    files.push('test/helpers/karma-init.js');
+  if (file) {
+    file = Array.isArray(file) ? ['test/pipeline_setup.js', ...file] : [file]
   }
+
+  var files = file ? ['test/test_deps.js', ...file, 'test/helpers/hookSetup.js'].flatMap(f => f) : ['test/test_index.js'];
 
   var config = {
     // base path that will be used to resolve all patterns (eg. files, exclude)
@@ -127,15 +125,15 @@ module.exports = function(codeCoverage, browserstack, watchMode, file, disableFe
     },
     // frameworks to use
     // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
-    frameworks: ['es5-shim', 'mocha', 'chai', 'sinon'],
+    frameworks: ['es5-shim', 'mocha', 'chai', 'sinon', 'webpack'],
 
-    files: files,
+    // test files should not be watched or they'll run twice after an update
+    // (they are still, in fact, watched through autoWatch: true)
+    files: files.map(fn => ({pattern: fn, watched: false, served: true, included: true})),
 
     // preprocess matching files before serving them to the browser
     // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
-    preprocessors: {
-      'test/test_index.js': ['webpack', 'sourcemap']
-    },
+    preprocessors: Object.fromEntries(files.map(f => [f, ['webpack', 'sourcemap']])),
 
     // web server port
     port: 9876,
@@ -148,7 +146,7 @@ module.exports = function(codeCoverage, browserstack, watchMode, file, disableFe
     logLevel: karmaConstants.LOG_INFO,
 
     // enable / disable watching file and executing tests whenever any file changes
-    autoWatch: true,
+    autoWatch: watchMode,
 
     reporters: ['mocha'],
 
@@ -174,15 +172,6 @@ module.exports = function(codeCoverage, browserstack, watchMode, file, disableFe
 
     plugins: plugins
   };
-
-  // To ensure that, we are able to run single spec file
-  // here we are adding preprocessors, when file is passed
-  if (file) {
-    config.files.forEach((file) => {
-      config.preprocessors[file] = ['webpack', 'sourcemap'];
-    });
-    delete config.preprocessors['test/test_index.js'];
-  }
 
   setReporters(config, codeCoverage, browserstack);
   setBrowsers(config, browserstack);
