+++ test/helpers/cookies.js
@@ -1,0 +1,5 @@
+export function clearAllCookies() {
+  document.cookie.split(';').forEach(function (c) {
+    document.cookie = c.replace(/^ +/, '').replace(/=.*/, '=;expires=' + new Date().toUTCString() + ';path=/');
+  });
+}
