--- test/helpers/karma-init.js
@@ -1,6 +1,0 @@
-(function (window) {
-  if (!window.parent.pbjsKarmaInitDone && window.location.pathname === '/context.html') {
-    window.parent.pbjsKarmaInitDone = true;
-    window.open('/debug.html', '_blank');
-  }
-})(window);
