+++ test/spec/modules/dsaControl_spec.js
@@ -108,6 +108,5 @@ describe('DSA transparency', () => {
         })
       })
     })
-    it('should accept bids regardless of dsa when "required" any other value')
   });
 });
