+++ test/spec/modules/euidIdSystem_spec.js
@@ -1,4 +1,4 @@
-import {coreStorage, init, setSubmoduleRegistry} from 'modules/userId/index.js';
+import {attachIdSystem, coreStorage, init, setSubmoduleRegistry} from 'modules/userId/index.js';
 import {config} from 'src/config.js';
 import {euidIdSubmodule} from 'modules/euidIdSystem.js';
 import 'modules/consentManagement.js';
@@ -8,6 +8,7 @@ import {apiHelpers, cookieHelpers, runAuction, setGdprApplies} from './uid2IdSys
 import {hook} from 'src/hook.js';
 import {uninstall as uninstallGdprEnforcement} from 'modules/gdprEnforcement.js';
 import {server} from 'test/mocks/xhr';
+import {createEidsArray} from '../../../modules/userId/eids.js';
 
 let expect = require('chai').expect;
 
@@ -161,4 +162,24 @@ describe('EUID module', function() {
       expectOptout(bid, optoutToken);
     });
   }
+
+  describe('eid', () => {
+    before(() => {
+      attachIdSystem(euidIdSubmodule);
+    });
+    it('euid', function() {
+      const userId = {
+        euid: {'id': 'Sample_AD_Token'}
+      };
+      const newEids = createEidsArray(userId);
+      expect(newEids.length).to.equal(1);
+      expect(newEids[0]).to.deep.equal({
+        source: 'euid.eu',
+        uids: [{
+          id: 'Sample_AD_Token',
+          atype: 3
+        }]
+      });
+    });
+  })
 });
