+++ test/spec/modules/tapadIdSystem_spec.js
@@ -2,6 +2,9 @@ import { tapadIdSubmodule, graphUrl } from 'modules/tapadIdSystem.js';
 import * as utils from 'src/utils.js';
 
 import { server } from 'test/mocks/xhr.js';
+import {createEidsArray} from '../../../modules/userId/eids.js';
+import {expect} from 'chai/index.mjs';
+import {attachIdSystem} from '../../../modules/userId/index.js';
 
 describe('TapadIdSystem', function () {
   describe('getId', function() {
@@ -62,4 +65,20 @@ describe('TapadIdSystem', function () {
       logMessageSpy.restore();
     });
   });
+  describe('eid', () => {
+    before(() => {
+      attachIdSystem(tapadIdSubmodule);
+    });
+    it('tapadId', function() {
+      const userId = {
+        tapadId: 'some-random-id-value'
+      };
+      const newEids = createEidsArray(userId);
+      expect(newEids.length).to.equal(1);
+      expect(newEids[0]).to.deep.equal({
+        source: 'tapad.com',
+        uids: [{id: 'some-random-id-value', atype: 1}]
+      });
+    });
+  })
 })
